Move QB modal types from querying into the store state types
