+++ frontend/src/metabase/query_builder/components/view/ViewHeader/ViewTitleHeader.tsx
@@ -4,8 +4,8 @@ import { useEffect } from "react";
 import { usePrevious } from "react-use";
 
 import { useToggle } from "metabase/common/hooks/use-toggle";
-import type { QueryModalType } from "metabase/querying/constants";
 import type { DatasetEditorTab, QueryBuilderMode } from "metabase/redux/store";
+import type { QueryModalType } from "metabase/redux/store/qb";
 import { Flex } from "metabase/ui";
 import * as Lib from "metabase-lib";
 import type Question from "metabase-lib/v1/Question";
