+++ frontend/src/metabase/query_builder/components/view/ViewHeader/components/QuestionActions/QuestionActions.tsx
@@ -10,9 +10,9 @@ import { UploadInput } from "metabase/common/components/upload";
 import { useRegisterShortcut } from "metabase/palette/hooks/useRegisterShortcut";
 import { runQuestionQuery } from "metabase/query_builder/actions";
 import { QuestionMoreActionsMenu } from "metabase/query_builder/components/view/ViewHeader/components/QuestionActions/QuestionMoreActionsMenu";
-import type { QueryModalType } from "metabase/querying/constants";
 import { useDispatch } from "metabase/redux";
 import type { DatasetEditorTab, QueryBuilderMode } from "metabase/redux/store";
+import type { QueryModalType } from "metabase/redux/store/qb";
 import { UploadMode } from "metabase/redux/store/upload";
 import { uploadFile } from "metabase/redux/uploads";
 import { Box, Divider, Icon, Menu } from "metabase/ui";
